修正

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: [:new, :edit]
+  before_action :authenticate_user!, only: [:new, :edit, :destroy]
   def index
     @items = Item.order('created_at DESC')
   end
@@ -37,7 +37,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if current_user.id == item.user_id
     redirect_to root_path
   end
